Do not fail Assimp build from source if warnings are emitted
Different compilers may emit different warnings when building Assimp.
While it's good for the upstream Assimp project to try to minimize
warnings by failing a build on a compiler warning by default, for us it
creates unnecessary failure modes that are very likely to be caught
upstream anyway. For example, GCC 13 as shipped on the latest Ubuntu
release has an alleged compiler bug that triggers a warning, and thus
fails the Assimp build, even though end-user functionality is not
affected.

Let's define the CMake variable documented upstream to not treat
warnings as errors, which should increase the resilience of our build
from source process. Users interested in fixing the root cause of the
warnings can contribute to assimp itself.