Use `libz-sys` to build and link to zlib and not reinvent wheels
Assimp allows building zlib from its vendored source, but that's not
necessary if zlib is provided in other ways, such as by means of the
`libz-sys` crate. `russimp-sys` does not need to reinvent the logic that
`libz-sys` handles well already.

As a side bonus, this allows users to statically link to zlib.